// src/app/api/dto/AdminCardsSetting.g.ts
/*tslint:disable*/

export interface AdminCardsSetting {
    maxTimeInCart: number;
    minTimeInCart: number;
    maxCardsInCart: number;
    maxAmountInCart: number;
    minimumRestForResale: number;
    mobileWarrantyDays: number;
    partnerWarrantyDays: number;
    bulkSellerWarrantyDays: number;
}